// src/app/core/models/water-quality.model.ts
export interface testing_points {
  id: string,
  organizationId:string,
  pointCode: string,
  pointName: string,
  pointType: PointType, 
  zoneId: string,
  streetId?: string, // Optional field for DOMICILIO type points
  locationDescription: string,
  coordinates: {
    latitude: number,
    longitude: number
  },
  status: Status,
  created_at: string
}

export enum PointType{
  RESERVORIO ='RESERVORIO',
  RED_DISTRIBUCION='RED_DISTRIBUCION',
  DOMICILIO='DOMICILIO'
}
// ...
export enum Status{
    ACTIVE ='ACTIVE',
    INACTIVE = 'INACTIVE'
}